+++ src/test/java/roomescape/acceptance/DatabaseConnectionTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -15,7 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
         properties = "spring.datasource.url=jdbc:h2:mem:database"
 )
 @Transactional
-public class MissionStep2Test {
+public class DatabaseConnectionTest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
