+++ src/test/java/roomescape/acceptance/LayerArchitectureTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -12,7 +12,7 @@ import roomescape.controller.ReservationController;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @Transactional
-public class MissionStep4Test {
+public class LayerArchitectureTest {
 
     @Autowired
     private ReservationController reservationController;
