+++ src/test/java/roomescape/config/TestTimeConfig.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.config;
 
 import java.time.Clock;
 import java.time.Instant;
