+++ src/test/java/roomescape/repository/JdbcThemeRepositoryTest.java
@@ -13,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.TestTimeConfig;
+import roomescape.config.TestTimeConfig;
 import roomescape.domain.Theme;
 
 @JdbcTest
